requirements.txt
# PDF işleme
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Excel işleme
openpyxl>=3.1.0
pandas>=2.0.0

# Word işleme
python-docx>=1.0.0

# PDF üretimi
reportlab>=4.0.0

# Görsel işleme
Pillow>=10.0.0

# Yapılandırma
PyYAML>=6.0.0

# CLI
rich>=13.0.0
questionary>=2.0.0

# Anthropic API
anthropic>=0.18.0

# ======== YENİ: v3.0 Modülleri ========

# RAG Sistemi (Vektör veritabanı - local embedding)
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Grafik Üretimi
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1

# Veri İşleme
numpy>=1.24.0

# ======== YENİ: v4.0 Web Araştırma ========

# Web araştırma
duckduckgo_search>=4.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0

# ======== YENİ: v5.0 Gelişmiş RAG ========

# Gelişmiş Embedding Modelleri
FlagEmbedding>=1.2.0

# Hybrid Search (BM25)
rank-bm25>=0.2.2

# Token Sayımı
tiktoken>=0.5.0

# Async İşlem
aiohttp>=3.9.0

# Keyword Extraction
scikit-learn>=1.3.0
